refactor(getPageSchema): rename file & remove unused export
